fix(trusty-common): split ETXTBSY retry into spawn_retry.rs (500-SLOC cap)
CI's ratcheted 500-line file-size cap flagged supervisor.rs at 527 SLOC
after the ETXTBSY retry was added inline. Split spawn_with_etxtbsy_retry
and the trusty-embedderd command construction into a new sibling module
(embedder_client/spawn_retry.rs), matching this file's own precedent of
splitting out supervisor_tests.rs for the same reason. supervisor.rs now
just calls spawn_retry::spawn_embedderd(binary_path, config) and is back
under the cap (500 SLOC via check_line_cap.sh).

Refs #3570
